implement authorized default in BaseProcessor, other processors can use default.
